Prints summary of augmentedRCBD object

print.augmentedRCBD prints to console the summary of an object of class augmentedRCBD including the augmented design details, ANOVA (Treatment adjusted), ANOVA (Block adjusted), treatment means, coefficient of variation, overall adjusted mean, critical differences and standard errors. The treatment/genotype groups along with the grouping method are also printed if they were computed.

Usage

# S3 method for class 'augmentedRCBD'
print(x, ...)

Arguments

x

An object of class augmentedRCBD.

...

Unused

See also