Analysis of Augmented Randomised Complete Block Design

augmentedRCBD is a function for analysis of variance of an augmented randomised block design (Federer, 1956; Federer, 1961) and the generation as well as comparison of the adjusted means of the treatments/genotypes.

augmentedRCBD(
  block,
  treatment,
  y,
  checks = NULL,
  method.comp = c("lsd", "tukey", "none"),
  alpha = 0.05,
  group = TRUE,
  console = TRUE,
  simplify = FALSE,
  truncate.means = TRUE
)

Arguments

block
Vector of blocks (as a factor).

treatment
Vector of treatments/genotypes (as a factor).

y
Numeric vector of response variable (Trait).

checks
Character vector of the checks present in treatment levels. If not specified, checks are inferred from the data on the basis of number of replications of treatments/genotypes.

method.comp
Method for comparison of treatments ("lsd" for least significant difference or "tukey" for Tukey's honest significant difference). If "none", no comparisons will be made, the ANOVA output will be given as a data frame and the adjusted means will be computed directly from treatment and block effects instead of using emmeans.

alpha
Type I error probability (Significance level) to be used for multiple comparisons.

group
If TRUE, genotypes will be grouped according to "method.comp". Default is TRUE.

console
If TRUE, output will be printed to console. Default is TRUE. Default is TRUE.

simplify
If TRUE, ANOVA output will be given as a data frame instead of a summary.aov object. Default is TRUE.

truncate.means
If TRUE, the negative adjusted means will be truncated to zero. Default is TRUE.

Value

A list of class augmentedRCBD containing the following components:

Details

Details of the augmented design used.

Means

A data frame with the "Means", "Block", "SE", "Mix", "Max" and "Adjusted Means" for each "Treatment".

ANOVA, Treatment Adjusted

An object of class summary.aov for ANOVA table with treatments adjusted.

ANOVA, Block Adjusted

An object of class summary.aov for ANOVA table with block adjusted.

Block effects

A vector of block effects.

Treatment effects

A vector of treatment effects.

Std. Errors

A data frame of standard error of difference between various combinations along with critical difference and tukey's honest significant difference (when method.comp = "tukey") at alpha.

Overall adjusted mean

Overall adjusted mean.

CV

Coefficient of variation.

Comparisons

A data frame of pairwise comparisons of treatments. This is computed only if argument group is TRUE

Groups

A data frame with compact letter display of pairwise comparisons of treatments. Means with at least one letter common are not significantly different statistically. This is computed only if argument group is TRUE

Details

This function borrows code from DAU.test function of agricolae package (de Mendiburu et al., 2016) as well as from Appendix VIII of Mathur et al., (2008).

Note

Data should preferably be balanced i.e. all the check genotypes should be present in all the blocks. If not, a warning is issued.

There should not be any missing values.

The number of test genotypes can vary within a block.

In case the large number of treatments or genotypes, it is advisable to avoid comparisons with the group = FALSE argument as it will be memory and processor intensive. Further it is advised to simplify output with simplify = TRUE in order to reduce output object size.

References

Federer WT (1956). “Augmented (or Hoonuiaku) designs.” The Hawaiian Planters' Record, LV(2), 191--208.

Federer WT (1956). “Augmented (or Hoonuiaku) Designs.” Technical Report BU-74-M, Cornell University, New York.

Federer WT (1961). “Augmented designs with one-way elimination of heterogeneity.” Biometrics, 17(3), 447--473.

Mathur PN, Muralidharan K, Parthasarathy VA, Batugal P, Bonnot F (2008). Data Analysis Manual for Coconut Researchers-Bioversity Technical Bulletin No. 14. Bioversity International. ISBN 978-92-9043-736-9.

de Mendiburu F (2015). agricolae: Statistical Procedures for Agricultural Research. R package version 1.2-8.

See also

Examples

# Example data
blk <- c(rep(1,7),rep(2,6),rep(3,7))
trt <- c(1, 2, 3, 4, 7, 11, 12, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 10)
y1 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77, 78, 78,
        70, 75, 74)
y2 <- c(258, 224, 238, 278, 347, 300, 289, 260, 220, 237, 227, 281, 311, 250,
        240, 268, 287, 226, 395, 450)
data <- data.frame(blk, trt, y1, y2)
# Convert block and treatment to factors
data$blk <- as.factor(data$blk)
data$trt <- as.factor(data$trt)
# Results for variable y1 (checks inferred)
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)

#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1       84.66667 3.844188 3  79  92       84.66667
#> 2         10     3 74.00000       NA 1  74  74       77.25000
#> 3         11     1 89.00000       NA 1  89  89       86.50000
#> 4         12     1 82.00000       NA 1  82  82       79.50000
#> 5          2       79.00000 1.154701 3  77  81       79.00000
#> 6          3       82.00000 2.645751 3  78  87       82.00000
#> 7          4       83.33333 3.929942 3  78  91       83.33333
#> 8          5     2 79.00000       NA 1  79  79       78.25000
#> 9          6     3 75.00000       NA 1  75  75       78.25000
#> 10         7     1 96.00000       NA 1  96  96       93.50000
#> 11         8     3 70.00000       NA 1  70  70       73.25000
#> 12         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 2          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 3          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 1          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2

# Results for variable y2 (checks inferred)
out2 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                     alpha = 0.05, group = TRUE, console = TRUE)

#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1       84.66667 3.844188 3  79  92       84.66667
#> 2         10     3 74.00000       NA 1  74  74       77.25000
#> 3         11     1 89.00000       NA 1  89  89       86.50000
#> 4         12     1 82.00000       NA 1  82  82       79.50000
#> 5          2       79.00000 1.154701 3  77  81       79.00000
#> 6          3       82.00000 2.645751 3  78  87       82.00000
#> 7          4       83.33333 3.929942 3  78  91       83.33333
#> 8          5     2 79.00000       NA 1  79  79       78.25000
#> 9          6     3 75.00000       NA 1  75  75       78.25000
#> 10         7     1 96.00000       NA 1  96  96       93.50000
#> 11         8     3 70.00000       NA 1  70  70       73.25000
#> 12         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 2          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 3          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 1          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2

# Results for variable y1 (checks specified)
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE,
                      checks = c("1", "2", "3", "4"))

#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1       84.66667 3.844188 3  79  92       84.66667
#> 2         10     3 74.00000       NA 1  74  74       77.25000
#> 3         11     1 89.00000       NA 1  89  89       86.50000
#> 4         12     1 82.00000       NA 1  82  82       79.50000
#> 5          2       79.00000 1.154701 3  77  81       79.00000
#> 6          3       82.00000 2.645751 3  78  87       82.00000
#> 7          4       83.33333 3.929942 3  78  91       83.33333
#> 8          5     2 79.00000       NA 1  79  79       78.25000
#> 9          6     3 75.00000       NA 1  75  75       78.25000
#> 10         7     1 96.00000       NA 1  96  96       93.50000
#> 11         8     3 70.00000       NA 1  70  70       73.25000
#> 12         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 2          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 3          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 1          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2

# Results for variable y2 (checks specified)
out2 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE,
                      checks = c("1", "2", "3", "4"))

#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1       84.66667 3.844188 3  79  92       84.66667
#> 2         10     3 74.00000       NA 1  74  74       77.25000
#> 3         11     1 89.00000       NA 1  89  89       86.50000
#> 4         12     1 82.00000       NA 1  82  82       79.50000
#> 5          2       79.00000 1.154701 3  77  81       79.00000
#> 6          3       82.00000 2.645751 3  78  87       82.00000
#> 7          4       83.33333 3.929942 3  78  91       83.33333
#> 8          5     2 79.00000       NA 1  79  79       78.25000
#> 9          6     3 75.00000       NA 1  75  75       78.25000
#> 10         7     1 96.00000       NA 1  96  96       93.50000
#> 11         8     3 70.00000       NA 1  70  70       73.25000
#> 12         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 2          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 3          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 1          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2

if (FALSE) {
# Error in case checks not replicated across all blocks
# Check 1 and 4 not replicated in all 3 blocks
trt <- c(1, 2, 3, 14, 7, 11, 12, 1, 2, 3, 4, 5, 9, 13, 2, 3, 4, 8, 6, 10)
data$trt <- as.factor(trt)
table(data$trt, data$blk)
# Results for variable y1 (checks specified)
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE,
                      checks = c("1", "2", "3", "4"))
}

# Warning in case test treatments are replicated
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)

#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1       84.66667 3.844188 3  79  92       84.66667
#> 2         10     3 74.00000       NA 1  74  74       77.25000
#> 3         11     1 89.00000       NA 1  89  89       86.50000
#> 4         12     1 82.00000       NA 1  82  82       79.50000
#> 5          2       79.00000 1.154701 3  77  81       79.00000
#> 6          3       82.00000 2.645751 3  78  87       82.00000
#> 7          4       83.33333 3.929942 3  78  91       83.33333
#> 8          5     2 79.00000       NA 1  79  79       78.25000
#> 9          6     3 75.00000       NA 1  75  75       78.25000
#> 10         7     1 96.00000       NA 1  96  96       93.50000
#> 11         8     3 70.00000       NA 1  70  70       73.25000
#> 12         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 2          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 3          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 1          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2

out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE,
                      checks = c("2", "3"))

#> Warning: Test treatments are replicated

#> 
#> Augmented Design Details
#> ========================
#>                                  
#> Number of blocks           "3"   
#> Number of treatments       "12"  
#> Number of check treatments "2"   
#> Number of test treatments  "10"  
#> Check treatments           "2, 3"
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    1   13.5   13.50   0.501 0.5058  
#>   Treatment: Test and Test vs. Check 10  271.6   27.16   1.007 0.5210  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              1   13.5   13.50   0.501  0.506
#>   Treatment: Test               9  560.7   62.31   2.310  0.160
#>   Treatment: Test vs. Check     1    1.4    1.42   0.053  0.826
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Treatment Means
#> ===============
#>    Treatment Block    Means       SE r Min Max Adjusted Means
#> 1          1     1 84.66667 3.844188 3  79  92       84.66667
#> 2          1     2 84.66667 3.844188 3  79  92       84.66667
#> 3          1     3 84.66667 3.844188 3  79  92       84.66667
#> 4         10     3 74.00000       NA 1  74  74       77.25000
#> 5         11     1 89.00000       NA 1  89  89       86.50000
#> 6         12     1 82.00000       NA 1  82  82       79.50000
#> 7          2       79.00000 1.154701 3  77  81       79.00000
#> 8          3       82.00000 2.645751 3  78  87       82.00000
#> 9          4     2 83.33333 3.929942 3  78  91       83.33333
#> 10         4     3 83.33333 3.929942 3  78  91       83.33333
#> 11         4     1 83.33333 3.929942 3  78  91       83.33333
#> 12         5     2 79.00000       NA 1  79  79       78.25000
#> 13         6     3 75.00000       NA 1  75  75       78.25000
#> 14         7     1 96.00000       NA 1  96  96       93.50000
#> 15         8     3 70.00000       NA 1  70  70       73.25000
#> 16         9     2 78.00000       NA 1  78  78       77.25000
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===================
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.995369 22.01088
#> A Test Treatment and a Control Treatment            7.344688 17.97180
#> 
#> Treatment Groups
#> ==================
#> 
#> Method : lsd
#> 
#>    Treatment Adjusted Means       SE df lower.CL  upper.CL Group
#> 8          8       73.25000 5.609598  6 59.52381  86.97619    1 
#> 9          9       77.25000 5.609598  6 63.52381  90.97619    12
#> 10        10       77.25000 5.609598  6 63.52381  90.97619    12
#> 6          6       78.25000 5.609598  6 64.52381  91.97619    12
#> 5          5       78.25000 5.609598  6 64.52381  91.97619    12
#> 1          2       79.00000 2.998456  6 71.66304  86.33696    12
#> 12        12       79.50000 5.609598  6 65.77381  93.22619    12
#> 2          3       82.00000 2.998456  6 74.66304  89.33696    12
#> 4          4       83.33333 2.998456  6 75.99637  90.67029    12
#> 3          1       84.66667 2.998456  6 77.32971  92.00363    12
#> 11        11       86.50000 5.609598  6 72.77381 100.22619    12
#> 7          7       93.50000 5.609598  6 79.77381 107.22619     2